morse-chat-be/requirements.txt
fastapi
uvicorn[standard]
pydantic
websockets
bcrypt
python-jose
passlib
